moved to linalg/utils
