+++ python/hail/docs/methods/misc.rst
@@ -11,9 +11,7 @@ Miscellaneous
     grep
     maximal_independent_set
     rename_duplicates
-    array_windows
 
 .. autofunction:: grep
 .. autofunction:: maximal_independent_set
 .. autofunction:: rename_duplicates
-.. autofunction:: array_windows
