+++ python/hail/linalg/__init__.py
@@ -1,3 +1,5 @@
 from .blockmatrix import BlockMatrix
+from . import utils as utils
 
-__all__ = ['BlockMatrix']
+__all__ = ['BlockMatrix',
+           'utils']
