+++ python/hail/linalg/utils/__init__.py
@@ -1,0 +1,4 @@
+from .misc import array_windows, locus_windows
+
+__all__ = ['array_windows',
+           'locus_windows']
