+++ python/hail/methods/__init__.py
@@ -8,7 +8,7 @@ from .statgen import linear_regression, logistic_regression, linear_mixed_regres
     hwe_normalized_pca, pc_relate, SplitMulti, filter_alleles, filter_alleles_hts, \
     split_multi_hts, balding_nichols_model, ld_prune
 from .qc import sample_qc, variant_qc, vep, concordance, nirvana, summarize_variants
-from .misc import rename_duplicates, maximal_independent_set, filter_intervals, window_by_locus, array_windows, locus_windows
+from .misc import rename_duplicates, maximal_independent_set, filter_intervals, window_by_locus
 
 __all__ = ['trio_matrix',
            'linear_regression',
@@ -58,7 +58,5 @@ __all__ = ['trio_matrix',
            'de_novo',
            'filter_alleles',
            'filter_alleles_hts',
-           'array_windows',
-           'locus_windows',
            'summarize_variants',
            ]
